Start of 1.15 port. Doesn't compile yet!

+++ src/main/java/me/desht/pneumaticcraft/common/progwidgets/IProgWidget.java
@@ -33,7 +33,7 @@ public interface IProgWidget extends IProgWidgetBase {
 
     ResourceLocation getTexture();
 
-    Pair<Double, Double> getMaxUV();
+    Pair<Float,Float> getMaxUV();
 
     int getTextureSize();
